change: export tests.common.tdi_base.Base to tests.common.loader.TestCase and relative changes
